requirements.txt
numpy
scipy
matplotlib
opencv-python
CairoSVG
pyzmq
protobuf
joblib
Pillow
svgfig
